LRU accepted.
I used dual way linked list along with HashMap. That will guarantee o(1)
insert instead of o(log(n)) in heap.

* Why dual way linked list?
When a inserted node is used, it need to get refreshed. And in order to
maintain the structure of the list, we need to know the node before and
after it so we can remove the target node without revisit the entire
list.

* What the structure of the node?
The node is a inner class of the class LinkedList, which itself is also
a inner class of LRU cache.

The node contain target value, target key, as well as node.before and
node.after.

We need to store its value to save another hashMap. The hashMap mapped
key with the target node. And when return node's value.

We need to store its key is because the "full condition". When the cache
is full, we need to delete the linkedlist's head as well as the
reference in the hashmap. So hashMap's size is same as the real size of
the cache. I cannot delete the reference in the hashmap without first
knowing its key.

These are two cases to delete a node in the linkedlist.
1. The
cache is full, its head is delete.
2. the node is over written.

For the
first case, we need its key because we want to delete its reference. For
the second case, we don't necessarily need to delete its key as it will
be over written due to the duplicate key issue.